+++ CHANGELOG.rst
@@ -15,6 +15,9 @@ Next Release (TBD)
   (`#351 <https://github.com/awslabs/chalice/pull/351>`__)
 * Fix vendor directory contents not being importable locally
   (`#350 <https://github.com/awslabs/chalice/pull/350>`__)
+* Add support for binary payloads
+  (`#352 <https://github.com/awslabs/chalice/pull/352>`__)
+
 
 0.8.2
 =====